+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This changelog will need to be split between individual packages
 ### Fixed
 
 - [@http/request/body-as-object] accept parameters in `content-type` header
+- [@http/route] in `byMethod()` a 405 response now includes the `Allow` header
 
 ## [0.29.0]